whiskey/requirements.txt
Django>=4.1.7
djangorestframework>=3.14.0
daphne>=4.0.0
